+++ pkg/index/inverted/inverted.go
@@ -35,7 +35,6 @@ import (
 	"go.uber.org/multierr"
 
 	"github.com/apache/skywalking-banyandb/api/common"
-	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
 	"github.com/apache/skywalking-banyandb/pkg/convert"
 	"github.com/apache/skywalking-banyandb/pkg/index"
@@ -63,13 +62,14 @@ var (
 )
 
 // Analyzers is a map that associates each IndexRule_Analyzer type with a corresponding Analyzer.
-var Analyzers map[databasev1.IndexRule_Analyzer]*analysis.Analyzer
+var Analyzers map[string]*analysis.Analyzer
 
 func init() {
-	Analyzers = map[databasev1.IndexRule_Analyzer]*analysis.Analyzer{
-		databasev1.IndexRule_ANALYZER_KEYWORD:  analyzer.NewKeywordAnalyzer(),
-		databasev1.IndexRule_ANALYZER_SIMPLE:   analyzer.NewSimpleAnalyzer(),
-		databasev1.IndexRule_ANALYZER_STANDARD: analyzer.NewStandardAnalyzer(),
+	Analyzers = map[string]*analysis.Analyzer{
+		index.AnalyzerKeyword:  analyzer.NewKeywordAnalyzer(),
+		index.AnalyzerSimple:   analyzer.NewSimpleAnalyzer(),
+		index.AnalyzerStandard: analyzer.NewStandardAnalyzer(),
+		index.AnalyzerURL:      newURLAnalyzer(),
 	}
 }
 
@@ -126,7 +126,7 @@ func (s *store) Batch(batch index.Batch) error {
 			if f.Store {
 				tf.StoreValue()
 			}
-			if f.Key.Analyzer != databasev1.IndexRule_ANALYZER_UNSPECIFIED {
+			if f.Key.Analyzer != index.AnalyzerUnspecified {
 				tf = tf.WithAnalyzer(Analyzers[f.Key.Analyzer])
 			}
 			doc.AddField(tf)
@@ -156,7 +156,7 @@ func NewStore(opts StoreOpts) (index.SeriesStore, error) {
 			WithPersisterNapTimeMSec(int(opts.BatchWaitSec * 1000))
 	}
 	config := bluge.DefaultConfigWithIndexConfig(indexConfig)
-	config.DefaultSearchAnalyzer = Analyzers[databasev1.IndexRule_ANALYZER_KEYWORD]
+	config.DefaultSearchAnalyzer = Analyzers[index.AnalyzerKeyword]
 	config.Logger = log.New(opts.Logger, opts.Logger.Module(), 0)
 	w, err := bluge.OpenWriter(config)
 	if err != nil {
@@ -282,15 +282,15 @@ func (s *store) MatchTerms(field index.Field) (list posting.List, err error) {
 	return list, err
 }
 
-func (s *store) Match(fieldKey index.FieldKey, matches []string) (posting.List, error) {
-	if len(matches) == 0 || fieldKey.Analyzer == databasev1.IndexRule_ANALYZER_UNSPECIFIED {
+func (s *store) Match(fieldKey index.FieldKey, matches []string, opts *modelv1.Condition_MatchOption) (posting.List, error) {
+	if len(matches) == 0 || fieldKey.Analyzer == index.AnalyzerUnspecified {
 		return roaring.DummyPostingList, nil
 	}
 	reader, err := s.writer.Reader()
 	if err != nil {
 		return nil, err
 	}
-	analyzer := Analyzers[fieldKey.Analyzer]
+	analyzer, operator := getMatchOptions(fieldKey.Analyzer, opts)
 	fk := fieldKey.Marshal()
 	query := bluge.NewBooleanQuery()
 	if fieldKey.HasSeriesID() {
@@ -298,7 +298,7 @@ func (s *store) Match(fieldKey index.FieldKey, matches []string) (posting.List, 
 	}
 	for _, m := range matches {
 		query.AddMust(bluge.NewMatchQuery(m).SetField(fk).
-			SetAnalyzer(analyzer))
+			SetAnalyzer(analyzer).SetOperator(operator))
 	}
 	documentMatchIterator, err := reader.Search(context.Background(), bluge.NewAllMatches(query))
 	if err != nil {
@@ -313,6 +313,22 @@ func (s *store) Match(fieldKey index.FieldKey, matches []string) (posting.List, 
 		list.Insert(iter.Val().DocID)
 	}
 	return list, err
+}
+
+func getMatchOptions(analyzerOnIndexRule string, opts *modelv1.Condition_MatchOption) (*analysis.Analyzer, bluge.MatchQueryOperator) {
+	analyzer := Analyzers[analyzerOnIndexRule]
+	operator := bluge.MatchQueryOperatorOr
+	if opts != nil {
+		if opts.Analyzer != index.AnalyzerUnspecified {
+			analyzer = Analyzers[opts.Analyzer]
+		}
+		if opts.Operator != modelv1.Condition_MatchOption_OPERATOR_UNSPECIFIED {
+			if opts.Operator == modelv1.Condition_MatchOption_OPERATOR_AND {
+				operator = bluge.MatchQueryOperatorAnd
+			}
+		}
+	}
+	return analyzer, bluge.MatchQueryOperator(operator)
 }
 
 func (s *store) Range(fieldKey index.FieldKey, opts index.RangeOpts) (list posting.List, err error) {