+++ banyand/measure/query.go
@@ -41,7 +41,8 @@ import (
 )
 
 const (
-	preloadSize = 100
+	preloadSize    = 100
+	checkDoneEvery = 128
 )
 
 // Query allow to retrieve measure data points.
@@ -252,12 +253,16 @@ func (s *measure) searchBlocks(ctx context.Context, result *queryResult, sids []
 	if tstIter.Error() != nil {
 		return fmt.Errorf("cannot init tstIter: %w", tstIter.Error())
 	}
+	var hit int
 	for tstIter.nextBlock() {
-		select {
-		case <-ctx.Done():
-			return errors.WithMessagef(ctx.Err(), "interrupt: scanned %d blocks, remained %d/%d parts to scan", len(result.data), len(tstIter.piHeap), len(tstIter.piPool))
-		default:
+		if hit%checkDoneEvery == 0 {
+			select {
+			case <-ctx.Done():
+				return errors.WithMessagef(ctx.Err(), "interrupt: scanned %d blocks, remained %d/%d parts to scan", len(result.data), len(tstIter.piHeap), len(tstIter.piPool))
+			default:
+			}
 		}
+		hit++
 		bc := generateBlockCursor()
 		p := tstIter.piHeap[0]
 		bc.init(p.p, p.curBlock, qo)
@@ -432,12 +437,20 @@ type queryResult struct {
 	data             []*blockCursor
 	snapshots        []*snapshot
 	segments         []storage.Segment[*tsTable, option]
+	hit              int
 	loaded           bool
 	orderByTS        bool
 	ascTS            bool
 }
 
 func (qr *queryResult) Pull() *model.MeasureResult {
+	select {
+	case <-qr.ctx.Done():
+		return &model.MeasureResult{
+			Error: errors.WithMessagef(qr.ctx.Err(), "interrupt: hit %d", qr.hit),
+		}
+	default:
+	}
 	if !qr.loaded {
 		if len(qr.data) == 0 {
 			return nil
@@ -446,6 +459,12 @@ func (qr *queryResult) Pull() *model.MeasureResult {
 		cursorChan := make(chan int, len(qr.data))
 		for i := 0; i < len(qr.data); i++ {
 			go func(i int) {
+				select {
+				case <-qr.ctx.Done():
+					cursorChan <- i
+					return
+				default:
+				}
 				tmpBlock := generateBlock()
 				defer releaseBlock(tmpBlock)
 				if !qr.data[i].loadData(tmpBlock) {
@@ -461,16 +480,17 @@ func (qr *queryResult) Pull() *model.MeasureResult {
 
 		blankCursorList := []int{}
 		for completed := 0; completed < len(qr.data); completed++ {
-			select {
-			case <-qr.ctx.Done():
-				return &model.MeasureResult{
-					Error: errors.WithMessagef(qr.ctx.Err(), "interrupt: loaded %d/%d cursors", completed, len(qr.data)),
-				}
-			case result := <-cursorChan:
-				if result != -1 {
-					blankCursorList = append(blankCursorList, result)
-				}
+			result := <-cursorChan
+			if result != -1 {
+				blankCursorList = append(blankCursorList, result)
 			}
+		}
+		select {
+		case <-qr.ctx.Done():
+			return &model.MeasureResult{
+				Error: errors.WithMessagef(qr.ctx.Err(), "interrupt: blank/total=%d/%d", len(blankCursorList), len(qr.data)),
+			}
+		default:
 		}
 		sort.Slice(blankCursorList, func(i, j int) bool {
 			return blankCursorList[i] > blankCursorList[j]