+++ CHANGES.md
@@ -46,6 +46,7 @@ Release Notes.
 - Fix panic when removing a expired segment.
 - Fix panic when reading a disorder block of measure. This block's versions are not sorted in descending order.
 - Fix the bug that the etcd client doesn't reconnect when facing the context timeout in the startup phase.
+- Fix the bug that the long running query doesn't stop when the context is canceled.
 
 ### Documentation