+++ CHANGES.md
@@ -47,6 +47,7 @@ Release Notes.
 - Fix panic when reading a disorder block of measure. This block's versions are not sorted in descending order.
 - Fix the bug that the etcd client doesn't reconnect when facing the context timeout in the startup phase.
 - Fix the bug that the long running query doesn't stop when the context is canceled.
+- Fix the bug that merge block with different tags or fields.
 
 ### Documentation